+++ DSS4-MS-MOVIE-V1/src/main/java/com/dss/dss4msmoviev1/service/MovieServiceImpl.java
@@ -1,32 +1,35 @@
 package com.dss.dss4msmoviev1.service;
 
 import com.dss.dss4msmoviev1.dto.MovieDTO;
+import com.dss.dss4msmoviev1.dto.UpdateMovieDTO;
 import com.dss.dss4msmoviev1.dto.util.MovieDTOMapper;
 import com.dss.dss4msmoviev1.entity.Movie;
+import com.dss.dss4msmoviev1.exception.CannotDeleteMovieException;
+import com.dss.dss4msmoviev1.exception.MovieNotFoundException;
 import com.dss.dss4msmoviev1.repository.MovieRepository;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.time.Year;
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
 public class MovieServiceImpl implements MovieService{
 
-    private static final Logger LOG =   LoggerFactory.getLogger(MovieServiceImpl.class);
     @Autowired
     private MovieRepository movieRepository;
+
+    private static final String NO_MOVIE_MESSAGE = "No movie found with id:";
+
 
     @Override
     public List<Movie> getAllMovies() {
         List<Movie> movieList = null;
         movieList = movieRepository.findAll();
+        if(movieList.isEmpty()){
+            throw new MovieNotFoundException("No movies found.");
+        }
         return movieList;
     }
 
@@ -34,11 +37,11 @@ public class MovieServiceImpl implements MovieService{
     public Movie getMovieById(int movieId) {
         Movie movieFound = null;
 
-        try{
-            Optional<Movie> movie = movieRepository.findById(movieId);
+        Optional<Movie> movie = movieRepository.findById(movieId);
+        if(!movie.isEmpty()) {
             movieFound = movie.get();
-        }catch (NoSuchElementException | NullPointerException e){
-            LOG.error(e.getMessage());
+        }else{
+            throw new MovieNotFoundException( NO_MOVIE_MESSAGE + movieId + ".");
         }
 
         return movieFound;
@@ -47,45 +50,45 @@ public class MovieServiceImpl implements MovieService{
     @Override
     public List<Movie> getMoviesByTitle(String movieTitle) {
         List<Movie> movieFound = null;
-        try{
-            movieFound = movieRepository.findByMovieTitle(movieTitle);
-            movieFound.size(); //for checking
-        }catch (NoSuchElementException | NullPointerException e){
-            LOG.error(e.getMessage());
+
+        movieFound = movieRepository.findByMovieTitle(movieTitle);
+
+        if(!movieFound.isEmpty()){
+            return movieFound;
+        }else{
+            throw new MovieNotFoundException("No movies found with title:" + movieTitle + ".");
         }
-        return movieFound;
+
     }
 
     @Override
     public List<Movie> getMoviesByYear(int movieYear) {
         List<Movie> movieFound = null;
-        try{
-            movieFound = movieRepository.findByMovieYear(movieYear);
-            movieFound.size(); //for checking
-        }catch (NoSuchElementException | NullPointerException e){
-            LOG.error(e.getMessage());
+        movieFound = movieRepository.findByMovieYear(movieYear);
+
+        if(!movieFound.isEmpty()){
+            return movieFound;
+        }else{
+            throw new MovieNotFoundException("No movies found with year:" + movieYear + ".");
         }
-        return movieFound;
     }
 
     @Override
     public List<Movie> getMoviesByActorId(int actorId) {
         List<Movie> movieFound = null;
-        try{
-            movieFound = movieRepository.findAllByActorsActorId(actorId);
-            movieFound.size(); //for checking
-        }catch (NoSuchElementException | NullPointerException e){
-            LOG.error(e.getMessage());
-        }
-        return movieFound;
-    }
+        movieFound = movieRepository.findAllByActorsActorId(actorId);
 
+        if(!movieFound.isEmpty()){
+            return movieFound;
+        }else{
+            throw new MovieNotFoundException("No movies found with actor ID:" + actorId + ".");
+        }
+
+    }
     @Override
     public Movie addMovie(MovieDTO movieDTO) {
-        MovieDTOMapper movieDTOMapper = new MovieDTOMapper();
-        Movie movie = movieDTOMapper.mapMovie(movieDTO);
-        Movie newMovie = movieRepository.save(movie);
-        return newMovie;
+        Movie movie = MovieDTOMapper.mapMovie(movieDTO);
+        return movieRepository.save(movie);
     }
 
     @Override
@@ -93,44 +96,42 @@ public class MovieServiceImpl implements MovieService{
         String responseMessage = null;
         Movie movieFound = null;
         int year = Year.now().getValue();
-        try {
-            Optional<Movie> movie = movieRepository.findById(movieId);
+
+        Optional<Movie> movie = movieRepository.findById(movieId);
+
+        if(!movie.isEmpty()) {
             movieFound = movie.get();
-            if((year - movieFound.getMovieYear()) >= 1){
+            if ((year - movieFound.getMovieYear()) >= 1) {
                 movieRepository.deleteById(movieId);
                 responseMessage = "Data successfully deleted.";
-            }else{
+                return responseMessage;
+            } else {
                 responseMessage = "Can't delete movie entry.";
+                throw new CannotDeleteMovieException(responseMessage);
             }
-        }catch(EmptyResultDataAccessException | NoSuchElementException e){
-            responseMessage = "No such data with id = " + movieId + ".";
-            LOG.error(e.getMessage());
+        }else{
+            throw new MovieNotFoundException(NO_MOVIE_MESSAGE + movieId + ".");
         }
-        return responseMessage;
+
     }
 
     @Override
-    public String updateMovie(int movieId, MovieDTO movieDTO) {
+    public String updateMovie(int movieId, UpdateMovieDTO movieDTO) {
         String responseMessage = null;
         Movie movieFound = null;
-        MovieDTOMapper movieDTOMapper = new MovieDTOMapper();
+        Optional<Movie> movie = movieRepository.findById(movieId);
 
-        try{
-            Optional<Movie> movie = movieRepository.findById(movieId);
+        if(!movie.isEmpty()) {
             movieFound = movie.get();
-            Movie movieUpdate = movieDTOMapper.mapMovie(movieDTO);
-            movieFound.setMovieCost(movieUpdate.getMovieCost());
-            movieFound.setMovieImage(movieUpdate.getMovieImage());
-            movieFound.setActors(movieUpdate.getActors());
-            movieFound.setMovieTitle(movieUpdate.getMovieTitle());
-            movieFound.setMovieYear(movieUpdate.getMovieYear());
+            movieFound.setMovieCost(movieDTO.getCost());
+            movieFound.setMovieImage(movieDTO.getImage());
             movieRepository.save(movieFound);
             responseMessage = "Data successfully updated.";
-
-        }catch (NoSuchElementException | NullPointerException e){
-            responseMessage = "No such data with id = " + movieId + ".";
-            LOG.error(e.getMessage());
+            return responseMessage;
+        }else{
+            throw new MovieNotFoundException(NO_MOVIE_MESSAGE + movieId + ".");
         }
-        return responseMessage;
+
+
     }
 }
